+++ server/apps/dcis/permissions/__init__.py
@@ -1,4 +1,5 @@
-from .sheet_permissions import AddChildRowDimension, ChangeSheet, ChangeValue, DeleteRowDimension, ViewDocument
+from .sheet_permissions import AddChildRowDimension, ChangeSheet, DeleteRowDimension, ViewDocument
+from .value_permissions import ChangeValue
 from .project_permissions import AddProject, ChangeProject, DeleteProject
 from .period_permissions import AddPeriod, ChangePeriod, DeletePeriod
 from .document_permissions import AddDocument, ChangeDocument, DeleteDocument, AddDocumentStatus, DeleteDocumentStatus
