+++ server/apps/dcis/schema/__init__.py
@@ -1,7 +1,11 @@
 import graphene
 
-from apps.dcis.schema.mutations import CellMutations, DocumentMutations, ProjectMutations, SheetMutations, \
-    PrivilegeMutations
+from apps.dcis.schema.mutations import CellMutations, \
+    DocumentMutations, \
+    ProjectMutations, \
+    SheetMutations, \
+    PrivilegeMutations, \
+    ValueMutations
 from apps.dcis.schema.queries import DocumentQueries, PrivilegeQueries, ProjectQueries, SheetQueries
 
 
@@ -23,6 +27,7 @@ class Mutation(
     DocumentMutations,
     SheetMutations,
     PrivilegeMutations,
+    ValueMutations,
     graphene.ObjectType,
 ):
     """Мутации приложения dcis."""
