+++ server/apps/dcis/schema/mutations/__init__.py
@@ -1,5 +1,6 @@
 from .cell_mutations import CellMutations
 from .document_mutations import DocumentMutations
+from .privilege_mutations import PrivilegeMutations
 from .project_mutations import ProjectMutations
 from .sheet_mutations import SheetMutations
-from .privilege_mutations import PrivilegeMutations
+from .value_mutations import ValueMutations
